rename flash -> common_options
The module does more than just flashing, so a more general name is warranted.
